added new ssr page that displays random pasta recipes
